Calculate and mail product discount rate

+++ user-service/src/main/java/com/bihju/domain/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String detailUrl;
     private double price;
     private double oldPrice;
-    private int discountPercent;
+    private double discountPercent;
     private long createTime;
     private long updateTime;
 }
